+++ src/views/UserEdit/UserEdit.js
@@ -29,6 +29,7 @@ import {
   deliveryFulfillmentValues,
 } from '../../constants';
 import { resourcesLoaded, showErrorCallout } from './UserEditHelpers';
+import withUserRoles from '../../components/Wrappers/withUserRoles';
 
 class UserEdit extends React.Component {
   static propTypes = {
@@ -46,6 +47,9 @@ class UserEdit extends React.Component {
     mutator: PropTypes.object,
     getProxies: PropTypes.func,
     getSponsors: PropTypes.func,
+    // assignedRoleIds, updateUserRoles comes from withUserRoles HOC
+    updateUserRoles: PropTypes.func,
+    assignedRoleIds: PropTypes.arrayOf(PropTypes.string),
   }
 
   static contextType = CalloutContext;
@@ -66,6 +70,7 @@ class UserEdit extends React.Component {
       getSponsors,
       getPreferredServicePoint,
       getUserServicePoints,
+      assignedRoleIds,
       match,
     } = this.props;
 
@@ -103,6 +108,7 @@ class UserEdit extends React.Component {
       proxies: getProxies(),
       sponsors: getSponsors(),
       servicePoints: getUserServicePoints(),
+      assignedRoleIds,
       requestPreferences: {
         ...initialFormValues.requestPreferences,
         ...get(this.props.resources, 'requestPreferences.records[0].requestPreferences[0]', {})
@@ -113,6 +119,7 @@ class UserEdit extends React.Component {
   getUserFormData() {
     const {
       resources,
+      assignedRoleIds
     } = this.props;
     const formData = getRecordObject(
       resources,
@@ -122,7 +129,7 @@ class UserEdit extends React.Component {
       'departments',
     );
 
-    return formData;
+    return { ...formData, assignedRoleIds };
   }
 
   createRequestPreferences = (requestPreferencesData, userId) => {
@@ -193,6 +200,7 @@ class UserEdit extends React.Component {
       updateProxies,
       updateSponsors,
       updateServicePoints,
+      updateUserRoles,
       mutator,
       history,
       resources,
@@ -230,6 +238,10 @@ class UserEdit extends React.Component {
 
     if (stripes.hasPerm('inventory-storage.service-points-users.item.post,inventory-storage.service-points-users.item.put')) {
       updateServicePoints(servicePoints, preferredServicePoint);
+    }
+
+    if (stripes.hasInterface('roles')) {
+      updateUserRoles(user.assignedRoleIds);
     }
 
     const data = omit(user, ['creds', 'proxies', 'sponsors', 'permissions', 'servicePoints', 'preferredServicePoint']);
@@ -414,9 +426,10 @@ class UserEdit extends React.Component {
         history={history}
         stripes={this.props.stripes}
         profilePictureConfig={profilePictureConfig}
+        assignedRoleIds={this.props.assignedRoleIds}
       />
     );
   }
 }
 
-export default withOkapiKy(UserEdit);
+export default withOkapiKy(withUserRoles(UserEdit));