+++ src/views/UserEdit/UserEdit.test.js
@@ -17,6 +17,10 @@ const userFormData = {
   },
 };
 
+jest.mock('../../components/Wrappers/withUserRoles', () => (Component) => {
+  return Component;
+});
+
 jest.mock('@folio/stripes/components', () => ({
   ...jest.requireActual('@folio/stripes/components'),
   LoadingView: () => 'LoadingView',
@@ -53,6 +57,7 @@ describe('UserEdit', () => {
   let props = {
     stripes: {
       hasPerm: jest.fn().mockReturnValue(true),
+      hasInterface: jest.fn().mockReturnValue(true),
       okapi: {
         tenant: 'tenantId',
       },
@@ -87,6 +92,7 @@ describe('UserEdit', () => {
     updateProxies: jest.fn(),
     updateSponsors: jest.fn(),
     updateServicePoints: jest.fn(),
+    updateUserRoles:jest.fn(),
     getUserServicePoints: jest.fn(),
     getPreferredServicePoint: jest.fn(),
     mutator: {