+++ src/views/UserEdit/UserForm.js
@@ -105,6 +105,7 @@ class UserForm extends React.Component {
     form: PropTypes.object, // provided by final-form
     intl: PropTypes.object,
     profilePictureConfig: PropTypes.object.isRequired,
+    assignedRoleIds: PropTypes.arrayOf(PropTypes.string),
   };
 
   static defaultProps = {
@@ -441,7 +442,11 @@ class UserForm extends React.Component {
                           form={form}
                           initialValues={initialValues}
                           setButtonRef={this.setButtonRef}
-                        /> : <EditUserRoles accordionId="userRoles" />
+                        /> : <EditUserRoles
+                          form={form}
+                          initialValues={initialValues}
+                          accordionId="userRoles"
+                        />
                       }
                       <EditServicePoints
                         accordionId="servicePoints"